Unified app entry points
