+++ react/features/app/components/App.web.js
@@ -1,3 +1,5 @@
+/* global APP */
+
 import { appInit } from '../actions';
 import { AbstractApp } from './AbstractApp';
 
@@ -44,6 +46,17 @@ export class App extends AbstractApp {
         super.componentWillMount(...args);
 
         this._getStore().dispatch(appInit());
+
+        // Stops collecting the logs and disposing the API when the user closes
+        // the page.
+        window.addEventListener('beforeunload', () => {
+            // Stop the LogCollector
+            if (APP.logCollectorStarted) {
+                APP.logCollector.stop();
+                APP.logCollectorStarted = false;
+            }
+            APP.API.dispose();
+        });
     }
 
     /**
