+++ react/features/app/components/index.js
@@ -1,1 +1,2 @@
 export * from './App';
+export { default as Root } from './Root';
