+++ react/features/base/react/index.js
@@ -1,5 +1,6 @@
 export * from './components';
 export * from './functions';
+export { default as AppRegistry } from './AppRegistry';
 export { default as Platform } from './Platform';
 export { default as RouteRegistry } from './RouteRegistry';
 export { default as Symbol } from './Symbol';
